move db schema to static module
